Moved checkSession from App.jsx
